Save expense using values typed by user

+++ app/src/main/java/com/victorbrndls/pfs/core/expense/dto/EditExpenseData.kt
@@ -1,10 +1,11 @@
 package com.victorbrndls.pfs.core.expense.dto
 
+import java.math.BigDecimal
 import java.util.*
 
 data class EditExpenseData(
     val id: Long?,
     val description: String,
     val date: Date,
-    val amount: Long
+    val amount: BigDecimal
 )
