+++ app/src/main/java/com/victorbrndls/pfs/core/expense/entity/Expense.kt
@@ -1,10 +1,11 @@
 package com.victorbrndls.pfs.core.expense.entity
 
+import java.math.BigDecimal
 import java.util.*
 
 data class Expense(
     val id: Long,
     val description: String,
     val date: Date,
-    val amount: Long
+    val amount: BigDecimal
 )
